Move spec_string logic out of style module
